+++ src/constants.js
@@ -28,10 +28,11 @@ CONSTANT(CALLBACK_PROMISE_OFFSET, 2);
 CONSTANT(CALLBACK_RECEIVER_OFFSET, 3);
 CONSTANT(CALLBACK_SIZE, 4);
 //Layout for ._bitField
-//[RR]RR GWFN CTBH RUDE LLLL LLLL LLLL LLLL
+//[RR]RO GWFN CTBH IUDE LLLL LLLL LLLL LLLL
 //[RR] = [Reserved] (Both bits are either on or off to represent
 //                    1 bit due to 31-bit integers in 32-bit v8)
 //R = [Reserved]
+//O = returnedNonUndefined
 //G = isAsyncGuaranteed
 //W = isFollowing (The promise that is being followed is not stored explicitly)
 //F = isFulfilled
@@ -40,13 +41,14 @@ CONSTANT(CALLBACK_SIZE, 4);
 //T = isFinal (used for .done() implementation)
 //I = isRejectionIgnored
 //B = isBound
-//R = isMigrated
+//I = isRejectionIgnored
 //H = isRejectionUnhandled
 //U = isUnhanldedRejectionNotified
 //D = isDisposable
 //E = isCancelled
 //L = Length, 16 bit unsigned
 CONSTANT(NO_STATE, 0x0|0);
+CONSTANT(RETURNED_NON_UNDEFINED, 0x10000000|0);
 CONSTANT(IS_ASYNC_GUARANTEED, 0x8000000|0);
 CONSTANT(IS_FOLLOWING, 0x4000000|0);
 CONSTANT(IS_FULFILLED, 0x2000000|0);
@@ -55,7 +57,7 @@ CONSTANT(IS_CANCELLABLE, 0x800000|0);
 CONSTANT(IS_FINAL, 0x400000|0);
 CONSTANT(IS_BOUND, 0x200000|0);
 CONSTANT(IS_REJECTION_UNHANDLED, 0x100000|0);
-CONSTANT(IS_REJECTION_IGNORED, 0x200000|0);
+CONSTANT(IS_REJECTION_IGNORED, 0x80000|0);
 CONSTANT(IS_UNHANDLED_REJECTION_NOTIFIED, 0x40000|0);
 CONSTANT(IS_DISPOSABLE, 0x20000|0);
 CONSTANT(IS_CANCELLED, 0x10000|0);