+++ src/debuggability.js
@@ -18,10 +18,15 @@ var debugging = !!(util.env("BLUEBIRD_DEBUG") != 0 &&
                         (__DEBUG__ ||
                          util.env("BLUEBIRD_DEBUG") ||
                          util.env("NODE_ENV") === "development"));
+
 var warnings = !!(util.env("BLUEBIRD_WARNINGS") != 0 &&
     (debugging || util.env("BLUEBIRD_WARNINGS")));
+
 var longStackTraces = !!(util.env("BLUEBIRD_LONG_STACK_TRACES") != 0 &&
     (debugging || util.env("BLUEBIRD_LONG_STACK_TRACES")));
+
+var wForgottenReturn = util.env("BLUEBIRD_W_FORGOTTEN_RETURN") != 0 &&
+    (warnings || !!util.env("BLUEBIRD_W_FORGOTTEN_RETURN"));
 
 Promise.prototype.suppressUnhandledRejections = function() {
     var target = this._target();
@@ -49,6 +54,14 @@ Promise.prototype._ensurePossibleRejectionHandled = function () {
 Promise.prototype._notifyUnhandledRejectionIsHandled = function () {
     fireRejectionEvent(REJECTION_HANDLED_EVENT,
                                   unhandledRejectionHandled, undefined, this);
+};
+
+Promise.prototype._setReturnedNonUndefined = function() {
+    this._bitField = this._bitField | RETURNED_NON_UNDEFINED;
+};
+
+Promise.prototype._returnedNonUndefined = function() {
+    return (this._bitField & RETURNED_NON_UNDEFINED) !== 0;
 };
 
 Promise.prototype._notifyUnhandledRejection = function () {
@@ -109,12 +122,25 @@ Promise.onUnhandledRejectionHandled = function (fn) {
                                  : undefined;
 };
 
+var disableLongStackTraces = function() {};
 Promise.longStackTraces = function () {
     if (async.haveItemsQueued() && !config.longStackTraces) {
         throw new Error(LONG_STACK_TRACES_ERROR);
     }
     if (!config.longStackTraces && longStackTracesIsSupported()) {
+        var Promise_captureStackTrace = Promise.prototype._captureStackTrace;
+        var Promise_attachExtraTrace = Promise.prototype._attachExtraTrace;
         config.longStackTraces = true;
+        disableLongStackTraces = function() {
+            if (async.haveItemsQueued() && !config.longStackTraces) {
+                throw new Error(LONG_STACK_TRACES_ERROR);
+            }
+            Promise.prototype._captureStackTrace = Promise_captureStackTrace;
+            Promise.prototype._attachExtraTrace = Promise_attachExtraTrace;
+            Context.deactivateLongStackTraces();
+            async.enableTrampoline();
+            config.longStackTraces = false;
+        };
         Promise.prototype._captureStackTrace = longStackTracesCaptureStackTrace;
         Promise.prototype._attachExtraTrace = longStackTracesAttachExtraTrace;
         Context.activateLongStackTraces();
@@ -128,11 +154,23 @@ Promise.hasLongStackTraces = function () {
 
 Promise.config = function(opts) {
     opts = Object(opts);
-    if ("longStackTraces" in opts && opts.longStackTraces) {
-        Promise.longStackTraces();
+    if ("longStackTraces" in opts) {
+        if (opts.longStackTraces) {
+            Promise.longStackTraces();
+        } else if (!opts.longStackTraces && Promise.hasLongStackTraces()) {
+            disableLongStackTraces();
+        }
     }
     if ("warnings" in opts) {
-        config.warnings = !!opts.warnings;
+        var warningsOption = opts.warnings;
+        config.warnings = !!warningsOption;
+        wForgottenReturn = config.warnings;
+
+        if (util.isObject(warningsOption)) {
+            if ("wForgottenReturn" in warningsOption) {
+                wForgottenReturn = !!warningsOption.wForgottenReturn;
+            }
+        }
     }
     if ("cancellation" in opts && opts.cancellation && !config.cancellation) {
         if (async.haveItemsQueued()) {
@@ -276,13 +314,15 @@ function longStackTracesAttachExtraTrace(error, ignoreSelf) {
     }
 }
 
-function checkForgottenReturns(returnValue, promiseCreated, name, promise) {
-    if (returnValue === undefined &&
-        promiseCreated !== null &&
-        config.longStackTraces &&
-        config.warnings) {
+function checkForgottenReturns(returnValue, promiseCreated, name, promise,
+                               parent) {
+    if (returnValue === undefined && promiseCreated !== null &&
+        wForgottenReturn) {
+        if (parent !== undefined && parent._returnedNonUndefined()) return;
+
+        if (name) name = name + " ";
         var msg = "a promise was created in a " + name +
-            " handler but was not returned from it";
+            "handler but was not returned from it";
         promise._warn(msg, true, promiseCreated);
     }
 }
@@ -735,40 +775,27 @@ var fireGlobalEvent = (function() {
             }
         };
     } else {
-        var customEventWorks = false;
-        var anyEventWorks = true;
-        try {
-            var ev = new self.CustomEvent("test");
-            customEventWorks = ev instanceof CustomEvent;
-        } catch (e) {}
-        if (!customEventWorks) {
-            try {
-                var event = document.createEvent("CustomEvent");
-                event.initCustomEvent("testingtheevent", false, true, {});
-                self.dispatchEvent(event);
-            } catch (e) {
-                anyEventWorks = false;
-            }
-        }
-        if (anyEventWorks) {
-            fireDomEvent = function(type, detail) {
-                var event;
-                // WebWorkers and Up-to-date browsers
-                if (customEventWorks) {
-                    event = new self.CustomEvent(type, {
-                        detail: detail,
-                        bubbles: false,
-                        cancelable: true
-                    });
-                // IE9
-                } else if (self.dispatchEvent) {
-                    event = document.createEvent("CustomEvent");
-                    event.initCustomEvent(type, false, true, detail);
-                }
+        var globalObject = typeof self !== "undefined" ? self :
+                     typeof window !== "undefined" ? window :
+                     typeof global !== "undefined" ? global :
+                     this !== undefined ? this : null;
 
-                return event ? !self.dispatchEvent(event) : false;
+        if (!globalObject) {
+            return function() {
+                return false;
             };
         }
+
+        try {
+            var event = document.createEvent("CustomEvent");
+            event.initCustomEvent("testingtheevent", false, true, {});
+            globalObject.dispatchEvent(event);
+            fireDomEvent = function(type, detail) {
+                var event = document.createEvent("CustomEvent");
+                event.initCustomEvent(type, false, true, detail);
+                return !globalObject.dispatchEvent(event);
+            };
+        } catch (e) {}
 
         var toWindowMethodNameMap = {};
         toWindowMethodNameMap[UNHANDLED_REJECTION_EVENT] = ("on" +
@@ -778,12 +805,12 @@ var fireGlobalEvent = (function() {
 
         return function(name, reason, promise) {
             var methodName = toWindowMethodNameMap[name];
-            var method = self[methodName];
+            var method = globalObject[methodName];
             if (!method) return false;
             if (name === REJECTION_HANDLED_EVENT) {
-                method.call(self, promise);
+                method.call(globalObject, promise);
             } else {
-                method.call(self, reason, promise);
+                method.call(globalObject, reason, promise);
             }
             return true;
         };